requirements.txt
urllib3>=1.22
pandas>=0.23.4
biopython>=1.70